Update patch set 13
Patch Set 13:

Stefan - just blank-out the version numbering but keep the history.
Add a non versioned history reccord to which we ammend along the B-release development. 
Just such we start going with the increments on the docs for the b-release.
We will take care of versioning and all that when we aproach the release.
Thx/Jonas
Versions, etc

Patch-set: 13