Add dependency checker tool
A simple script, check_dependencies.sh, has been put in the build
root. Given a file as argument, it will attempt to verify that all
URIs within said file are accessible. If not, an error message will be
presented and the script will exit with return code 1.

The script is added to the ODL plugin makefile, and used to verify the
Ubuntu and CentOS dependencies within.

Change-Id: I4bac9eb5e383629794f6f72baa5472f258e35bef
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>