Update of endpoint addresses
As https endpoints introduced in Fuel 7.0 has caused some
issues for the automated test setup, this reverts back to
the "non https" behavior as in 6.1 for both Horizon and
the endpoints, disabling https.

This is a *momentary* solution, we need to figure out how
to setup the testing infrastructure with secure endpoints
enabled!

Change-Id: Ida61c2fd0834861b26ab64afec87b2bfbd921e0e
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>