Support for selecting plugins during build
In order to cut the build time for unofficial builds, this change
supports setting the environment variable "BUILD_FUEL_PLUGINS" prior
to building.

Only the plugin targets from fuel/build/f_isoroot/Makefile that are
specified will then be built. In order to completely disable the
building of plugins, the environment variable is set to " ".

When using this functionality, the resulting iso file will be
prepended with the prefix "unofficial-" to clearly indiciate that
this is not a full build.

Change-Id: Ib0a914ed6577f54f31b056a752691316302081dc
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>