Update patch set 2
Patch Set 2:

Jonas: Good idea, will do!

Ruijing: This is not meant for the Gerrit builds, just for "developer builds" so that it is possible to more quickly build and test for instance (which is my use case) Fuel uplifts and changes without having to rebuild all plugins.

Patch-set: 2
Label: Verified=0