A simple Fuel@OPNFV deployment configuration framework
This patch provides a simple framework for configuring a fuel@OPNFV
deployment in terms of:
- HW and POD environment configuration
- OpenStack environment configuration
- OPNFV vallue added features enablement and configuration (plug-ins)

It does not completely provide separation of concerns between HW and POD admins (Pharos), CI/CD pipeline owners (Releng and Functest), installers and collaborative projects. But it is a simple and prgmatic first step towards that goal.

Change-Id: Id182e86ecd19e9bed59e772524934fcbaef7cd56
Signed-off-by: Jonas Bjurel <jonas.bjurel@ericsson.com>