Reaping improvements for Fuel 7
In order to download the deployment information, the node id
must be explicitly specified.

The fuel setting commmand returns "ha_compact" as the mode for
a cluster, but Fuel does itself not understand this when changing
the settings - it needs to be named "ha".

Added shebang for reap.py and deploy.py.

Change-Id: I3ecacb83dc44454b90dedc98104658a16926dc1f
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>