Fix for cache handling
Introduced a separate function "getcommitid" in the cache.sh
- a.k.a. $(CACHETOOL) - so that given a repo and a tag, branch
or commitid the corresponding commitid is returned.

The same principle should be used in all caching functions where
not a strict tag or head is specified (as git ls-remote will not
be able to figure out the commit id in those cases).

Change-Id: I3540dca7ab408c872eaabdc68d4058f9d6d7c45f
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>