Align all plugins with new method for cache id
The cache tool is now responsible for returning the commit
ID of a branch, tag or commit. This change was first
introduced in the bgpvpn plugin and now the rest of the
plugins with a need for this are also using the same method.

Without this, a plugin pointing to a commit ID on master
would fail to build once the head of master is updated to
point to a later commit...

Change-Id: I7c32d5b41871741717ae42c1334e8557f0b0bedd
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>