Update patch set 3
Patch Set 3:

Failing due to lack of zip on the build machine.

Jonas, is zip something you guys can just install on the ci host? 

I can rewrite it to use tar but will take a while to verify as I'll also have to change the use of unzip in the manifests and test a redeploy. 

/B

Patch-set: 3