Proposal for new templates directory structure
The current directory structure is named after the env type; virtual or
hardware. Since we are deploying fuel in multiple different labs (Ericsson,
Intel, LF), in multiple different setups (HA, NOHA), and on virtual and bare
metal, it is a bit troublesome to work with current setup due to labs having
different setups.

This proposal introduces lab name as a top level directory and hardware and
virtual environment directories underneath.

Please note that the hardware_environment templates may not be switched to use
of this new directory structure due to release activities and can be kept as they
are.

Change-Id: I1ba1f1a7dd3951f4ce403ad2efde6a76130d533e
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>