Fuel 8 uplift
Introducing Fuel 8 with associated changes:

Improvements in usage of cgroups - they are now bind mounted from
the host which solved a Docker-in-Docker problem.

The second layer Docker is now using devicemapper as the backend
which seems to failure in some host environments when running yum
in Docker (see https://github.com/docker/docker/issues/6980 for
background).

The cache tool (cache.mk) has a new argument "getbiweek" which is
used by the top Makefile when building the Fuel ISO. This makes sure
that there's a fresh build every second week even if the cache is not
purged automatically based on its meta data. This is something we
should add everywhere (in a separate change set).

And obviously, Fuel is now building from stable/8.0 upstream.

Change-Id: I7e2ee737fd3bedc75f58539fbc3c5fd1fa08cfa1
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>