Update patch set 2
Patch Set 2: Code-Review+2

We may need to swap aufs for devmapper as backend as a general recommendation (at least needed in the second layer of Docker), I'll add that as a separate change in that case.

It is needed in the *second* layer of Docker for some platforms at least.

Patch-set: 2
Label: Code-Review=+2