Update patch set 1
Patch Set 1:

> +1 Looks good to me, once this gets merged I will do needed updates
 > of the deployment scenarios.

Ok so I will not modify scenarios in this change. One thing which I'm not sure how to proceed is to pointing to specific version of ODL. I used automatedweeklyreleases-1097 version but it was removed so I switched to one of daily snapshot [1] but don't know how long it will be available.

[1] https://github.com/openstack/fuel-plugin-opendaylight/blob/master/pre_build_hook#L10

Patch-set: 1