Update patch set 1
Patch Set 1: Code-Review+1

Hey there.

if you arent happy with this, what about adding the "rebuild" like into the ks.cfg file as a final section in %post - so that we have an option to do A or B (and can document this so that if people want to use native names then they simple can comment out / modify the ks.cfg"?

I mean, if you are really looking for another way..

Patch-set: 1