Update patch set 1
Patch Set 1:

hmm - if i understand correctly, either way if we have to do the "rebuild" we will get hit by the time penalty right?

could this be passed at the grub boot line (splash screen) as a parameter at the start maybe - so that it overrides and is in effect right from the boot of the FUEL node (from ISO?).

Patch-set: 1