Update patch set 1
Patch Set 1:

If I remember the logic here, .cacheid will be called first if we're running a cached build and since we're in Docker, there won't be a repo in there. The reason for having this typeof check in the $(ISOCACHE) target is just so that we don't pull again if .cacheid has been run.

There's no harm in this, but with the current logic it should probably not be needed, but I may be wrong! :)

Patch-set: 1