Update patch set 1
Patch Set 1:

So Michal, it is apparent that it happens/and that this change is needed:-)
Though I want to understand why it is not happening with CI, it is likely because CI runs with build.sh and you run Make all (or similar). Stefan - maybe you can explain this before we merge this patch.
And clearly this happends after the build is successfull.
BR/Jonas

Patch-set: 1