Update patch set 2
Patch Set 2: Code-Review+2

Looks good to me..

did a spot check on 60% of the files and see the addition that was made..

Im assumign that for the .rst the change in commenting syntax is due to requirement in the tool? (docs/index.rst for example).

Besides that - looks good.. send me the info about this tool you used (instead of a one-liner ;P)>

Patch-set: 2
Reviewer: Gerrit User 403 <403@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2