Update patch set 1
Patch Set 1: Code-Review+2

removal of old stuff is easy to +2 :) - as you said, needs to have a build tested against and deployed to ensure that there are no hooks left to this info - that is have a CI run on the patch set maybe?

thx for the work on this.
D

Patch-set: 1
Label: Code-Review=+2