Update patch set 1
Patch Set 1:

I wonder if we actually need to keep local copy of mos-ubuntu. By default main mos repository points to http://10.20.0.2:8080/liberty-8.0/ubuntu/x86_64 so this packages are already there. As far as I see mos8.0-updates and so on are empty, so probably there are not needed for deployment. If we then only keep packages from general ubuntu repositories and remove definitions of additional mos-* repos then we will reduce size of the ISO. I will check this today.

Patch-set: 1