Update patch set 1
Patch Set 1:

So we need this type of logic for the repo build to succeed as we are pinning down the exact versions in the b-release branch. This should work (I hope) both for master and in conjunction with https://gerrit.opnfv.org/gerrit/#/c/9471/ also for stable/b (and we need to cherry pick this there as well).

Patch-set: 1
Reviewer: Gerrit User 203 <203@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0