Update patch set 2
Patch Set 2:

(3 comments)

Do you want to have this script included on Fuel ISO? If yes then it would be better to move this to build directory and create simple makefile which will ensure that this script will be available on Fuel Master. It also give you a chance to define which packages you want to have to be included in local mirrors.
Depending on external resource is what we trying avoid, but as you said it should be resolved by creating plugin. Speaking of this, Daniel and me also think about creating tacker plugin for fuel, so maybe we can combine our efforts to make this plugin for SR1?

Patch-set: 2