Update patch set 1
Patch Set 1:

(1 comment)

Do you plan to cherry-pick this to brahmaputra branch? If not then ok, but in the other case I'm afraid that we can't use snapshot version inside brahmaputra release. We had similar discussion in case of patches required for bgpvpn, which were not a part of ODL SR1. Team decided that we should not relay on snapshot version of ODL, because it is no longer available after  60 days. Because of that Nikolas prepared a patch which is applied on the top of SR1 release: https://github.com/openstack/fuel-plugin-opendaylight/commit/b876df2345bf8c785dceea74f4c271fdfcc9563c

Patch-set: 1
Label: Verified=0