deploy.py: add a time stamp to the new ISO image
Add a time stamp to the newly create ISO image. This is to prevent name
collisions with other ISO images from earlier deploy jobs.

Ideally this will only be a temporary need, and releng will provide with
unique IDs for each job that we can use to this effect.

[ Alexandru Avadanii ]
Only signing for upstreaming purposes.
I remember Josep was not that convinced that this patch should be
upstreamed, but I think it's worth a look.

WARN: This change leads to the creation of multiple ISO files, and
relies on `virsh undefine` code to remove old ISOs, which recently
proved NOT to work as expected in Fuel@OPNFV.
However, `virsh undefine` works nicely in Armband, where deploys
are ran as non-root, so it should be merged once Fuel@OPNFV switches
to non-root deploys.

Change-Id: Ie98835a026dddd95a2f67d798613f8a9b226e8ca
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>