Update patch set 9
Patch Set 9: Code-Review+1

I think we need this functionality enabled as soon as possible since it is pretty troublesome to login to the nodes and look at the logs manually. Especially in an environment that is pretty diverse such as ours.

I say we focus on the actual change and have the other improvements done in separate patches. Please remember that we need changes reflected on CI as well so delaying this will delay the CI-side of things too.

Patch-set: 9
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1