Update patch set 1
Patch Set 1:

Hi Szilard,
we could definitely use the function interface in exec_cmd function instead of popen objects. Then to be able to catch any error messages, we would need to pipe stderr into stdout. I don't know if this is desirable (it changes the current API), but definitely doable.

I also realize that with the current implementation, if the command fails, the return string will only be the output of stdout, not stderr, which most likely has no indication of why the command failed... someone should fix it... wink wink :)

Patch-set: 1