Update patch set 1
Patch Set 1:

Yes I still needed it because I rely on this functionality in other patches of this series, including the restructuring of the ipmi class.

I admit the implementation of the retry loop is pretty dumb. As Peter pointed out, not checking why the command failed is certainly a concern, but if we care we will just not ask the function to blindly retry another time, we should manually check the output and take a decision. Also the use cases are arguably limited, here I was mainly mimicking the behavior in the ipmi class...

However... the current implementation does not provide the output of stderr to the user, so checking the return values will not give much information today. This should be addressed, but probably in another patch.

Patch-set: 1