Update patch set 1
Patch Set 1:

I'd propose one out of two things to make this a bit more generic.

Either go with Dan's proposal to inherit the settings from the host, which I think in general would be the right way,

or doing a similar structure for the virtual templates as we're doing for the Fuel configurations, let's separate this out on a per-lab type of config.

Patch-set: 1