Update patch set 1
Patch Set 1:

I like both suggestions and what Stefan proposed was the one in my mind.

Purpose of this patch was actually to verify Fuel9 dea/dha uplifts Jonas did and at the same time highlight this fact.

We have this problem already today; for pods with centos jumphost, we have separate fuel.xml. We can perhaps take this as an activity to overhaul template directory structure and usage.

https://gerrit.opnfv.org/gerrit/gitweb?p=fuel.git;a=tree;f=deploy/templates

Patch-set: 1