Update patch set 1
Patch Set 1:

Question Stefan - 

For the case of being able to support the rollout of Ostack Centors from FUEL then that would be easy yes +1 - since we just do a deploy,reap and move on.  

What isnt clear to me is if the requirement here is that the following:

FUEL VM runs on Centos Host.

Cheers,
D

Patch-set: 1