Update patch set 1
Patch Set 1:

Hey guys.. 

see the email thread, but i think that long term (and extending the idea of container) we should look to start imaging the JUMPHOST as well - why not run a vJUMPHOST with a VFUEL and then vOstack?
its only one more linux bridge if you think about it  and imaging the JUMPHOST gives nice isolated way for installer to be "unique" while having a comming approach from the "BARE METAL" up.

Daniel

Patch-set: 1