Update patch set 2
Patch Set 2: Code-Review-1

First off, we're adding an additional network here.

What we've had up until now is this:

fuel1: Admin (PXE) native, Managenent on VLAN
fuel2: Storage
fuel3: Private
fuel4: Public

With this change we'll have physical interfaces for all (and thus one additional bridge):

admin: Admin (PXE)
api: Management
storage: Storage
tenant: Private
external: Public

Having everything non-tagged would make sniffing for trouble shooting easier maybe, so that could be a possible advantage, although it requires us to change the DEA for all virtual pipelines (OPNFV CI, the devel-pipeline). Personally I don't care either way, the "fuel1 .. fuel4" mapping is etched in my brain circuitry, but if this makes it easier for newcomers then why not.

But I don't like the bridge names, as they don't reflect the Fuel naming, which is important in order to make this logical and easy for people to use (see the paragraph above). Also I'd like to see a "fuel-" prefix in the naming still, so it's clear for which VMs these bridges are setup. My suggestion would be:

fuel-admin: Admin (PXE)
fuel-mgmt: Management
fuel-storage: Storage
fuel-private: Private
fuel-public: Public

So I'm in no way opposing the change (although we need to sync it with corresponding DEA changes at the same time) - the -1 reflects the naming only.

Then, some more technical issues: I would expect the bridge setup for ha and noha to be identical as HA does not refer to network redundancy for the virtual setup. They're not though.

admin.xml (fuel-admin): Different IP addresses for the bridge in the HA and non-HA variants, also a NAT clause in one but not the other.

external-xml (fuel-public): An IP address in one, but not in the other, also a NAT clause in one but not the other.

Patch-set: 2
Reviewer: Gerrit User 203 <203@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1