Update patch set 1
Patch Set 1:

I don't think this patch is too big an issue for jenkins jobs, as
long as releng/jjb/fuel/fuel-deploy.sh is not changed. In
fuel-deploy.sh the bridge is set by default to pxebr.

However, this will affect people that manually run fuel/ci/deploy.sh.

I want to suggest two options to make the transition to using
"admin" as the default bridge:

1) Give the heads up, and push the change. I am pretty sure people
will fix their environment quite quickly if they haven't already ;-)

2) Do it in phases:

2.1) Add to this patch a modification to fuel/ci/deploy.sh to set
"pxebr" as the default bridge if -B is not given.
This way people that have not transitioned won't notice the change,
but those that have will have to explicitly set -B admin.

2.2) Modify the jenkins deploy script releng/jjb/fuel/fuel-deploy.sh
to accept an environment variable (BRIDGE or DEFAULT_BRIDGE) to
define the bridge to use for the deploy job. If the variable is not
set, the script will keep using pxebr as it does today.
Those PODs that have made the transition will set the corresponding
environment variable to "admin" in their corresponding section of the
slave-params.yml file.

2.3) At a previously announced date all PODs will have to have made
the transition, at which point the "pxebr" default bridge in
fuel/ci/deploy.sh will be removed.
Notice that those PODs that have not yet transitioned will STILL be
able to deploy normally by setting the environment variable for their
jenkins job as in 2).

I could help doing 2.1) and the first part of 2.2) if needed.

Patch-set: 1