repo mirror: Allow multi-arch local mirrors
Original patch from Stan, signed by me for upstreaming.

This patch allows specifying multiple architectures via UBUNTU_ARCH in
form of a list of space separated architectures. The first architecture
in the list is considered primary and will be used for building all the
deb packages by fuel-main. Additional architectures are added to allow
targets of other architectures to use the mirror.

NOTE: this imposes a requirement that all packages built are arch
independent (which is true so far).

[ Alexandru Avadanii ]
Reworked for applying on top of Fuel@OPNFV.
Pass UBUNTU_ARCH to Docker containers.

Change-Id: Ibc4ae03223545383813ac58ec34d44c1f97b4694
Signed-off-by: Stanislaw Kardach <stanislaw.kardach@cavium.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>