Update patch set 1
Patch Set 1: Code-Review+1

Hey there.

Looks ok to me, im wondering if there isnt a way of finding if the service is up in a more authoritative way, rather and a TIMEOUT value  - just wondering if on other systems 5 minutes will be "enough" and we have to meet this again?

Patch-set: 1