Update patch set 4
Patch Set 4:

Hi Szilard, I would definitely like these two patches to be merged, this would mean less patches to maintain in the armband repo! ;-) And there shoud not be any change in behavior from the current implementation, so this is backwards compatible with current setups.

However, if this is an inconvenience for fuel@opnfv, we (armband) can also wait.

Patch-set: 4