Aligning machine type over Ubuntu and Centos
We have previously used "pc-1.0" as the qemu-kvm machine definition
for virtual Fuel deploys. On Ubuntu this points to:

pc-1.0               Standard PC (i440FX + PIIX, 1996)

On CentOS 7 this machine definition does not exist, but there's a
match in "pc":

pc                   RHEL 7.0.0 PC (i440FX + PIIX, 1996) (alias of
pc-i440fx-rhel7.0.0

So, finally, pc is also in Ubuntu:

pc                   Ubuntu 14.04 PC (i440FX + PIIX, 1996) (alias of
pc-i440fx-trusty)

This change is about exploring whether moving to "pc" will make the
auto deploy work on both platforms with the "pc" machine type.

What still needs to be done in CentOS is to create a symlink from
/usr/libexec/qemu-kvm and /usr/bin/kvm though.

Change-Id: I1cce52a21075950cb59b75709b12d724d759e3e6
Signed-off-by: Stefan K. Berg <stefan.k.berg@ericsson.com>