Scenario/pod/lab for minimal local virtual deploy.
Existing local/virtual_kvm deploy didn't work out of the box for me so
this is a totally stripped down configuration (two nodes, no ceph for instance).

The lab dir structure seems to have been modified lately (local is gone
and devel-pipeline is in) so hopefully there is still a place for this
change.

Change-Id: I0ed1c6b4dcf59bd87a062d046a641aa245b3187b
Signed-off-by: Billy O'Mahony <billy.o.mahony@intel.com>