Update patch set 1
Patch Set 1:

> Could it be that on some hardware this will not run? or is it
 > hardware independet?

You are asking about kvm and ovs scenario functionality or VM? For example kvm will not be able to start when constant_tsc flag is not present, which is included in Intel processors since Pentium 4 if I recall correctly, though it my not work on AMD processors. If you are asking about VM it is a job o libvirt to discover host cpu model, and usually should not prevent vm start. But it is not ideal and may lead to some problems which libvirt documentation mention: https://libvirt.org/formatdomain.html#elementsCPU

Patch-set: 1
Label: Verified=0