Update patch set 1
Patch Set 1:

> (1 comment)

Hi Nikolas, actually no-ha scenarios are also tested by CI, but are run on virt env. Difference between no-ha an ha scenarios already exist and I'm not sure if we should worry about it. I agree that if it is possible we can add more resources to vms.

Patch-set: 1