Update patch set 1
Patch Set 1:

It is ok for me to track master branch of the plugin in master branch of fuel@opnfv. But I see one problem, currently any changes which we want to introduce to stable branch needs to go through master branch, there will be a problem when we will want to update plugin in stable/colorado. I'm not saying that I agree with above workflow especially in context of such change, just want highlight once more that requirement that everything needs to pass through master may not always work well.

Patch-set: 1
Label: Verified=0