Update patch set 2
Patch Set 2:

Just to comment on fuseiso and the *build* process - this should not be an issue for our non-Ubuntu friends as the build is run in the scope of the build Docker container.

I'm just thinking that for now perhaps the deploy code should potentially run this step thru a container as well to get around the dependency issue?

Patch-set: 2
Reviewer: Gerrit User 203 <203@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Workflow=0